pyproject.toml
[tool.poetry]
name = "botto"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
python-dotenv = "^0.18.0"
Pillow = "^8.3.0"
asyncpg = "^0.23.0"
"aiodog.py" = "^0.2.0"
more-itertools = "^8.8.0"
jishaku = {git = "https://github.com/Gorialis/jishaku"}

[tool.poetry.dev-dependencies]
taskipy = "^1.8.1"

[tool.taskipy.tasks]
start = "python -m bot"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
